+++ Tortuga.Chain/Tortuga.Chain.PostgreSql/PostgreSql/PostgreSqlMetadataCache.cs
@@ -787,42 +787,15 @@ WHERE (constrained_schema = '{tableName.Schema}' AND constrained_table = '{table
 
 	ColumnMetadataCollection<NpgsqlDbType> GetColumns(PostgreSqlObjectName tableName, NpgsqlConnection connection)
 	{
-		string columnSql;
-
-		if (ServerVersion.Major < 10) //no attidentity
-		{
-			columnSql =
-			@"
-SELECT att.attname as column_name,
-	   t.typname as data_type,
-	   pk.contype as is_primary_key,
-	   att.attnotnull as not_null,
-	   null as is_identity,
-	   format_type(att.atttypid, att.atttypmod) as data_type_full
-FROM pg_class as c
-JOIN pg_namespace as ns on ns.oid=c.relnamespace
-JOIN pg_attribute as att on c.oid=att.attrelid AND
-							att.attnum>0
-JOIN pg_type as t on t.oid=att.atttypid
-LEFT JOIN (SELECT cnst.conrelid,
-				  cnst.conkey,
-				  cnst.contype
-		   FROM pg_constraint as cnst
-		   WHERE cnst.contype='p') pk ON att.attnum=ANY(pk.conkey) AND
-										 pk.conrelid=c.oid
-WHERE c.relname ILIKE @Name AND
-	  ns.nspname ILIKE @Schema;";
-		}
-		else
-		{
-			columnSql =
+		const string columnSql =
 @"
 SELECT att.attname as column_name,
 	   t.typname as data_type,
 	   pk.contype as is_primary_key,
 	   att.attnotnull as not_null,
 	   att.attidentity as is_identity,
-	   format_type(att.atttypid, att.atttypmod) as data_type_full
+	   format_type(att.atttypid, att.atttypmod) as data_type_full,
+	   COL_DESCRIPTION(CONCAT(ns.nspname, '.', c.relname)::regclass, att.attnum) as description
 FROM pg_class as c
 JOIN pg_namespace as ns on ns.oid=c.relnamespace
 JOIN pg_attribute as att on c.oid=att.attrelid AND
@@ -836,7 +809,6 @@ LEFT JOIN (SELECT cnst.conrelid,
 										 pk.conrelid=c.oid
 WHERE c.relname ILIKE @Name AND
 	  ns.nspname ILIKE @Schema;";
-		}
 
 		var columns = new List<ColumnMetadata<NpgsqlDbType>>();
 		var sequenceColumns = GetSequenceColumns(tableName);
@@ -851,6 +823,7 @@ WHERE c.relname ILIKE @Name AND
 				{
 					var name = reader.GetString("column_name");
 					var typeName = reader.GetString("data_type");
+					var description = reader.GetStringOrNull("description");
 					var isPrimary = !reader.IsDBNull("is_primary_key");
 
 					var identity_type = char.ToUpperInvariant(reader.GetCharOrNull("is_identity") ?? ' ');
@@ -896,7 +869,7 @@ WHERE c.relname ILIKE @Name AND
 							break;
 					}
 
-					columns.Add(new ColumnMetadata<NpgsqlDbType>(name, false, isPrimary, isIdentity, typeName, dbType, QuoteColumnName(name), isNullable, maxLength, precision, scale, fullTypeName, ToClrType(typeName, isNullable, maxLength)));
+					columns.Add(new ColumnMetadata<NpgsqlDbType>(name, false, isPrimary, isIdentity, typeName, dbType, QuoteColumnName(name), isNullable, maxLength, precision, scale, fullTypeName, ToClrType(typeName, isNullable, maxLength)) { Description = description });
 				}
 			}
 		}
@@ -1139,15 +1112,16 @@ where s.relkind='S' and d.deptype='a'";
 	public TableOrViewMetadata<PostgreSqlObjectName, NpgsqlDbType> GetTableOrViewInternal(PostgreSqlObjectName tableName)
 	{
 		const string TableSql =
-			@"SELECT
-				table_schema as schemaname,
-				table_name as tablename,
-				table_type as type
-				FROM information_schema.tables
-				WHERE table_schema ILIKE @Schema AND
-					  table_name ILIKE @Name AND
-					  (table_type='BASE TABLE' OR
-					   table_type='VIEW');";
+@"SELECT
+	table_schema as schemaname,
+	table_name as tablename,
+	table_type as type,
+	OBJ_DESCRIPTION(CONCAT(table_schema, '.', table_name)::regclass) as description
+FROM information_schema.tables
+WHERE table_schema ILIKE @Schema AND
+		table_name ILIKE @Name AND
+		(table_type='BASE TABLE' OR
+		table_type='VIEW');";
 
 		using (var con = new NpgsqlConnection(m_ConnectionBuilder.ConnectionString))
 		{
@@ -1158,6 +1132,7 @@ where s.relkind='S' and d.deptype='a'";
 				string actualSchema;
 				string actualTableName;
 				bool isTable;
+				string? description;
 
 				using (var cmd = new NpgsqlCommand(TableSql, con))
 				{
@@ -1171,12 +1146,13 @@ where s.relkind='S' and d.deptype='a'";
 						actualTableName = reader.GetString("tablename");
 						var type = reader.GetString("type");
 						isTable = type.Equals("BASE TABLE", StringComparison.Ordinal);
+						description = reader.GetStringOrNull("description");
 					}
 				}
 
 				var actualName = new PostgreSqlObjectName(actualSchema, actualTableName);
 				var columns = GetColumns(actualName, con);
-				return new TableOrViewMetadata<PostgreSqlObjectName, NpgsqlDbType>(this, actualName, isTable, columns);
+				return new TableOrViewMetadata<PostgreSqlObjectName, NpgsqlDbType>(this, actualName, isTable, columns) { Description = description };
 			}
 		}